linsos tangents refactor
