+++ triples/core/__init__.py
@@ -1,13 +1,9 @@
 from .sum_of_squares import (
-    sum_of_squares, sum_of_squares_multiple, sum_of_square, sum_of_square_multiple,
-    METHOD_ORDER, DEFAULT_CONFIGS
+    sum_of_squares, sum_of_squares_multiple, METHOD_ORDER, DEFAULT_CONFIGS
 )
 
 from .linsos import (
-    LinearSOS,
-    root_tangents,
-    LinearBasis, LinearBasisTangent,
-    SolutionLinear
+    LinearSOS, LinearBasis, LinearBasisTangent, SolutionLinear
 )
 
 from .structsos import (
@@ -27,10 +23,8 @@ from .sdpsos import (
 )
 
 __all__ = [
-    'sum_of_squares', 'sum_of_squares_multiple', 'sum_of_square', 'sum_of_square_multiple',
-    'METHOD_ORDER', 'DEFAULT_CONFIGS',
-    'LinearSOS', 'root_tangents', 'LinearBasis', 'LinearBasisTangent',
-    'SolutionLinear',
+    'sum_of_squares', 'sum_of_squares_multiple', 'METHOD_ORDER', 'DEFAULT_CONFIGS',
+    'LinearSOS', 'LinearBasis', 'LinearBasisTangent', 'SolutionLinear',
     'StructuralSOS', 'SolutionStructural',
     'prove_univariate', 'prove_univariate_interval',
     'pqr_sym', 'pqr_cyc', 'pqr_ker',
