+++ triples/core/linsos/__init__.py
@@ -1,13 +1,10 @@
 from .linsos import LinearSOS
-from .tangents import root_tangents
 from .basis import (
     LinearBasis, LinearBasisTangent, LinearBasisTangentEven
 )
 from .solution import SolutionLinear
 
 __all__ = [
-    'LinearSOS',
-    'root_tangents',
-    'LinearBasis', 'LinearBasisTangent', 'LinearBasisTangentEven',
+    'LinearSOS', 'LinearBasis', 'LinearBasisTangent', 'LinearBasisTangentEven',
     'SolutionLinear'
 ]
