+++ triples/utils/monomials.py
@@ -3,7 +3,7 @@ from warnings import warn
 
 import numpy as np
 import sympy as sp
-from sympy.polys.domains import ZZ, QQ
+from sympy import Poly, ZZ, QQ
 from sympy.polys.polyclasses import DMP
 from sympy.combinatorics import Permutation, PermutationGroup, CyclicGroup, SymmetricGroup
 
@@ -43,8 +43,8 @@ def generate_partitions(nvars: int, degree: int, equal: bool = False) -> List[Tu
     # dict_monoms = {t: i for i, t in enumerate(inv_monoms)}
     # return dict_monoms, inv_monoms
 
-def _poly_rep(poly: Union[sp.Poly, DMP]) -> DMP:
-    return poly.rep if isinstance(poly, sp.Poly) else poly
+def _poly_rep(poly: Union[Poly, DMP]) -> DMP:
+    return poly.rep if isinstance(poly, Poly) else poly
 
 class MonomialManager():
     """
@@ -221,7 +221,7 @@ class MonomialManager():
             return False
         return True
 
-    def _arraylize_list(self, poly: Union[sp.Poly, DMP], expand_cyc: bool = False) -> List:
+    def _arraylize_list(self, poly: Union[Poly, DMP], expand_cyc: bool = False) -> List:
         rep = _poly_rep(poly)
         self._assert_equal_nvars(rep.lev + 1)
         degree = rep.total_degree()
@@ -240,7 +240,7 @@ class MonomialManager():
                         array[v] += coeff
         return array
 
-    def arraylize_np(self, poly: Union[sp.Poly, DMP], expand_cyc: bool = False) -> np.ndarray:
+    def arraylize_np(self, poly: Union[Poly, DMP], expand_cyc: bool = False) -> np.ndarray:
         """
         Return the vector representation of the polynomial in numpy array.
         """
@@ -256,7 +256,7 @@ class MonomialManager():
                 vec = [int(v) for v in vec]
         return np.array(vec).astype(np.float64)
 
-    def arraylize_sp(self, poly: Union[sp.Poly, DMP], expand_cyc: bool = False) -> sp.Matrix:
+    def arraylize_sp(self, poly: Union[Poly, DMP], expand_cyc: bool = False) -> sp.Matrix:
         """
         Return the vector representation of the polynomial in sympy matrix (column vector).
         """
@@ -270,17 +270,42 @@ class MonomialManager():
             vec = [to_sympy(v) for v in vec]
             return sp.Matrix(vec)
 
-    def invarraylize(self, array: Union[List, np.ndarray, sp.Matrix], gens: List[sp.Symbol], degree: int) -> sp.Poly:
+    def invarraylize(self, array: Union[List, np.ndarray, sp.Matrix], gens: List[sp.Symbol], degree: int) -> Poly:
         """
         Reverse the arraylize_np function to get the polynomial from its vector representation.
         """
         self._assert_equal_nvars(gens)
         inv_monoms = self.inv_monoms(degree)
         terms_dict = {}
+        permute = self.permute if (not self.is_trivial) else lambda x: (x,)
+        if SDM is not None and isinstance(array, sp.MatrixBase):
+            rep = array._rep.rep.to_sdm()
+            domain = rep.domain
+            zero = domain.zero
+            # be careful to handle column / row vectors
+            if rep.shape[0] == 1:
+                rep_list = list(rep.get(0, dict()).items())
+            elif rep.shape[1] == 1:
+                rep_list = [(i, v.get(0, zero)) for i, v in rep.items()]
+            elif rep.shape[0] == 0 or rep.shape[1] == 0:
+                return Poly(0, *gens)
+            else:
+                raise ValueError(f"Array must be a vector, but received shape {array.shape}.")
+            
+            for i, z in rep_list:
+                if z != zero:
+                    monom = inv_monoms[i]
+                    for monom2 in permute(monom):
+                        terms_dict[monom2] = z
+            rep = DMP.from_dict(terms_dict, len(gens)-1, domain)
+            return Poly.new(rep, *gens)
+
+        elif isinstance(array, np.ndarray):
+            array = array.tolist()
         for coeff, monom in zip(array, inv_monoms):
-            for monom2 in self.permute(monom):
+            for monom2 in permute(monom):
                 terms_dict[monom2] = coeff
-        return sp.Poly(terms_dict, gens)
+        return Poly(terms_dict, gens)
 
     def cyclic_sum(self, expr: sp.Expr, gens: List[sp.Symbol]) -> sp.Expr:
         """
@@ -330,7 +355,7 @@ def _parse_options(nvars, **options) -> MonomialManager:
     raise ValueError(f"Invalid symmetry type {type(symmetry)}. Expected MonomialManager or PermutationGroup.")
 
 
-def arraylize_np(poly: Union[sp.Poly, DMP], expand_cyc: bool = False, **options) -> np.ndarray:
+def arraylize_np(poly: Union[Poly, DMP], expand_cyc: bool = False, **options) -> np.ndarray:
     """
     Convert a sympy polynomial to a numpy vector of coefficients.
 
@@ -382,11 +407,11 @@ def arraylize_np(poly: Union[sp.Poly, DMP], expand_cyc: bool = False, **options)
     ----------
     arraylize_sp, invarraylize, generate_monoms
     """
-    nvars = (poly.rep if isinstance(poly, sp.Poly) else poly).lev + 1
+    nvars = (poly.rep if isinstance(poly, Poly) else poly).lev + 1
     option = _parse_options(nvars, **options)
     return option.arraylize_np(poly, expand_cyc = expand_cyc)
 
-def arraylize_sp(poly: Union[sp.Poly, DMP], expand_cyc: bool = False, **options) -> sp.Matrix:
+def arraylize_sp(poly: Union[Poly, DMP], expand_cyc: bool = False, **options) -> sp.Matrix:
     """
     Convert a sympy polynomial to a sympy vector of coefficients.
 
@@ -438,11 +463,11 @@ def arraylize_sp(poly: Union[sp.Poly, DMP], expand_cyc: bool = False, **options)
     ----------
     arraylize_np, invarraylize, generate_monoms
     """
-    nvars = (poly.rep if isinstance(poly, sp.Poly) else poly).lev + 1
+    nvars = (poly.rep if isinstance(poly, Poly) else poly).lev + 1
     option = _parse_options(nvars, **options)
     return option.arraylize_sp(poly, expand_cyc = expand_cyc)
 
-def invarraylize(array: Union[List, np.ndarray, sp.Matrix], gens: List[sp.Symbol], degree: int, **options) -> sp.Poly:
+def invarraylize(array: Union[List, np.ndarray, sp.Matrix], gens: List[sp.Symbol], degree: int, **options) -> Poly:
     """
     Convert a vector representation of polynomial back to the sympy polynomial.
